# swagger.json
{
  "openapi": "3.1.0",
  "info": {
    "version": "2.0.0",
    "title": "DREAM Lab API",
    "description": "Endpoints para la API de DREAM Lab. Puede encontrar más información en [GitHub](https://github.com/CodeCraft-Solutions-DREAM-Lab).",
    "contact": {
      "name": "CodeCraft Solutions",
      "url": "https://github.com/CodeCraft-Solutions-DREAM-Lab"
    }
  },
  "servers": [
    {
      "url": "http://localhost:3000/",
      "description": "Local server"
    },
    {
      "url": "https://dreamlab-api.azurewebsites.net/",
      "description": "Live server"
    }
  ],
  "paths": {
    "/auth/usuario": {
      "post": {
        "tags": [
          "Auth"
        ],
        "summary": "Autentica un usuario",
        "description": "Autentica un usuario",
        "responses": {
          "200": {
            "description": "OK",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "jwt": {
                      "type": "string",
                      "example": "eyJhbGciOiJIUzI1NiIsInR5cCI6IkpXVCJ9.eyJ1c3VhcmlvIjoiYTAxMTc3NzY3IiwiaWF0IjoxNzE1NzUzNzQzfQ.ml-vMvWq5X8_FdILT9YIPv0oPc9Wlvj3f_N4VhHCAZA"
                    },
                    "rol": {
                      "type": "string",
                      "example": "Regular"
                    }
                  }
                }
              }
            }
          },
          "401": {
            "description": "Unauthorized",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object"
                }
              }
            }
          },
          "404": {
            "description": "Not Found",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object"
                }
              }
            }
          }
        },
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "type": "object",
                "properties": {
                  "usuario": {
                    "type": "string",
                    "example": "test"
                  },
                  "contrasena": {
                    "type": "string",
                    "example": "test"
                  },
                  "origen": {
                    "type": "string",
                    "example": "qr"
                  },
                  "tagId": {
                    "type": "string",
                    "example": "test"
                  }
                }
              }
            }
          }
        }
      }
    },
...
    "/reservaciones/usuario/{id}": {
      "get": {
        "tags": [
          "Reservaciones"
        ],
        "summary": "Obtiene las reservaciones de un usuario",
        "description": "Obtiene las reservaciones de un usuario",
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string"
            },
            "description": "ID del usuario"
          }
        ],
        "responses": {
          "200": {
            "description": "OK",
            "content": {
              "application/json": {
                "schema": {
                  "type": "array",
                  "items": {
                    "type": "object",
                    "properties": {
                      "idReservacion": {
                        "type": "integer",
                        "example": 1
                      },
                      "idUsuario": {
                        "type": "string",
                        "example": "a01177767"
                      },
                      "idSala": {
                        "type": "integer",
                        "example": 2
                      },
                      "idExperiencia": {
                        "type": "integer",
                        "example": 2
                      },
                      "idMesa": {
                        "type": "integer",
                        "example": 2
                      },
                      "estatus": {
                        "type": "integer",
                        "example": 3
                      },
                      "estatusMateriales": {
                        "type": "integer",
                        "example": 1
                      },
                      "horaInicio": {
                        "type": "string",
                        "format": "date-time",
                        "example": "1970-01-01T12:00:00.000Z"
                      },
                      "duracion": {
                        "type": "integer",
                        "example": 3
                      },
                      "fecha": {
                        "type": "string",
                        "format": "date-time",
                        "example": "2024-01-05T00:00:00.000Z"
                      },
                      "numPersonas": {
                        "type": "integer",
                        "example": 3
                      },
                      "asistencia": {
                        "type": "null"
                      },
                      "nombreAlterno": {
                        "type": "null"
                      }
                    }
                  }
                }
              }
            }
          },
          "404": {
            "description": "Not Found"
          },
          "500": {
            "description": "Error",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "error": {
                      "type": "string"
                    }
                  }
                }
              }
            }
          }
        }
      }
    },
...
    "/usuarios/": {
      "get": {
        "tags": [
          "Usuarios"
        ],
        "summary": "Obtiene todos los usuarios",
        "description": "Obtiene todos los usuarios",
        "responses": {
          "200": {
            "description": "OK",
            "content": {
              "application/json": {
                "schema": {
                  "type": "array",
                  "items": {
                    "type": "object",
                    "properties": {
                      "idUsuario": {
                        "type": "string"
                      },
                      "nombre": {
                        "type": "string"
                      },
                      "apellidoP": {
                        "type": "string"
                      },
                      "apellidoM": {
                        "type": "string"
                      },
                      "tipo": {
                        "type": "string"
                      },
                      "prioridad": {
                        "type": "integer"
                      },
                      "logroPrincipal": {
                        "type": "integer"
                      },
                      "colorPreferido": {
                        "type": "string"
                      }
                    }
                  }
                }
              }
            }
          },
          "500": {
            "description": "Error",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "error": {
                      "type": "string"
                    }
                  }
                }
              }
            }
          }
        }
      }
    },
    "/usuarios/{idUsuario}": {
      "put": {
        "tags": [
          "Usuarios"
        ],
        "summary": "Modifica elementos de un usuario por su id.",
        "description": "Modifica a un usuario usando su id",
        "parameters": [
          {
            "name": "idUsuario",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string"
            },
            "description": "id del usuario"
          }
        ],
        "responses": {
          "200": {
            "description": "OK",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "rowsAffected": {
                      "type": "integer"
                    }
                  }
                }
              }
            }
          },
          "500": {
            "description": "Error",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "error": {
                      "type": "string"
                    }
                  }
                }
              }
            }
          }
        },
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "type": "object",
                "properties": {
                  "valor": {
                    "type": "integer"
                  }
                }
              }
            }
          }
        }
      }
    },
    "/usuarios/nombreUsuario/{tagID}": {
      "get": {
        "tags": [
          "Usuarios"
        ],
        "summary": "Obtener nombre de usuario por Tag ID",
        "description": "Obtiene el nombre de usuario asociado a un tag ID específico.",
        "parameters": [
          {
            "name": "tagID",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string"
            },
            "description": "ID del tag del usuario"
          }
        ],
        "responses": {
          "200": {
            "description": "OK",
            "content": {
              "application/json": {
                "schema": {
                  "type": "array",
                  "items": {
                    "type": "object",
                    "properties": {
                      "nombreUsuario": {
                        "type": "string"
                      }
                    }
                  }
                }
              }
            }
          },
          "500": {
            "description": "Error",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "error": {
                      "type": "string"
                    }
                  }
                }
              }
            }
          }
        }
      }
    },
    "/usuarios/cambiarPrioridad": {
      "post": {
        "tags": [
          "Usuarios"
        ],
        "summary": "Cambia los puntos de prioridad de un usuario",
        "description": "Cambia los puntos de prioridad de un usuario, lo guarda en el historial de cambios y envía un correo al usuario",
        "responses": {
          "200": {
            "description": "OK",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "message": {
                      "type": "string",
                      "example": "Prioridad updated"
                    }
                  }
                }
              }
            }
          },
          "400": {
            "description": "Faltan datos",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "message": {
                      "type": "string",
                      "example": "idUsuario, puntos and motivo are required"
                    }
                  }
                }
              }
            }
          },
          "500": {
            "description": "Error",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "error": {
                      "type": "string"
                    }
                  }
                }
              }
            }
          }
        },
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "type": "object",
                "properties": {
                  "idUsuario": {
                    "type": "string",
                    "example": "a00833173"
                  },
                  "puntos": {
                    "type": "integer",
                    "example": 10
                  },
                  "motivo": {
                    "type": "string"
                  }
                }
              }
            }
          }
        }
      }
    },
...
  }
}